convert: Normalize and extend wind/solar conversion arguments
Both now take a either a config dictionary describing the turbine or
panel or the basename of a reatlas config file for the turbine/panel
argument. Also the aggregation process can now be configured to
rescale the output per_unit, return the no_of_units per cell and the
capacity_factor, itself.