datasets: Use generators to allow for lazy xarray computation
prepare functions now yield a generator of datasets, when the generator is
exhausted the file is closed and memory is freed afterwards. The new pattern is
simpler and avoids having to load the dataset fully into memory as a preparation
to using dask, as necessary for SARAH.

Before the data handle is freed the generator has to be exhausted! See receive
context manager in sarah.py.