Update config system to use a object rather than a module.
The config system now uses a config object, located in the module "config".
The parent class is atlite.config.Config, access to the configuration is now achieved via atlite.config.config .
Getters and setters are implemented, the callback function for update hooks is mostly working, except for when the object atlite.config.config is overwritten directly with a new Config object.